backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
anthropic
python-dotenv
edge-tts
